Delete memory manager dead code.
There have existed two paths in the memory manager for
a few months -- the old path which uniformly distributes
all memory across all visible renderers, and the new path
which takes into account the needs of each renderer.

The transition was made from the old policy to the new
policy quite a while ago, so we can safely get rid of
the old policy.

BUG=150883

Review URL: https://chromiumcodereview.appspot.com/12475002

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@187678 4ff67af0-8c30-449e-8e8b-ad334ec8d88c