Eliminate redundant pref save in ~DownloadPrefs()
During destruction, DownloadPrefs attempts to save the internal
list of file types that should be opened automatically. However,
this list is also saved each time it is modified. Therefore the
attempt to save on destruction is redundant.

In addition, setting preferences during destruction interacts
poorly with the fast-shutdown pathway since it will cause
notifications to be dispatched for the changed preferences to be
dispatched during profile shutdown.

BUG=180442


Review URL: https://chromiumcodereview.appspot.com/12631006

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@187699 4ff67af0-8c30-449e-8e8b-ad334ec8d88c