Add per-profile disk caching of complied GPU shaders.
This CL adds a per-profile disk cache for any shaders that are
complied while using the profile. When the profile is first opened
the shaders will be loaded from disk and used to pre-populate the
GPU memory shader cache.

The disk cache takes the load time for From Dust from ~30 seconds
to ~18 seconds on my Linux machine for any loads after the first.

BUG=166763

Review URL: https://chromiumcodereview.appspot.com/12036056

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@187704 4ff67af0-8c30-449e-8e8b-ad334ec8d88c