Revert 186583 > Disabling ANOTHER flaky FullscreenControllerInteractiveTest: EscapingMouseLock. > > BUG=146006 > TBR=yzshen > > Review URL: https://codereview.chromium.org/12564002
The test was disabled due to the test_server, which has
had recent work done (http://crbug.com/96594) to reduce
flakiness and increase debuggability of why it sometimes
fails to start.

This test, EscapingMouseLock, does not appear to have any
individual contribution to the flakiness, as when it does
fail it is on the first line starting the test server.

Review URL: https://chromiumcodereview.appspot.com/12754007

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@187759 4ff67af0-8c30-449e-8e8b-ad334ec8d88c