Introduce DriveFileSystemGetter for DriveURLRequestJob.
This CL is preparation to write tests for DriveURLRequestJob.
It removes the void* for profile id from DriveURLRequestJob by replacing it
to DriveFileSystemGetter callback as an injecting point.
Along with the change, this CL introduces some thin wrapper of XxxOnUIThread
methods, which run on IO thread and sends a task to run XxxOnUIThread with
wrapping the given callbacks to run them on IO thread again. These methods are
for simplifying the deep nested function calls in an expression.

BUG=159383
TEST=Ran unit_tests and manually tried drive: scheme.


Review URL: https://chromiumcodereview.appspot.com/12446005

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@187836 4ff67af0-8c30-449e-8e8b-ad334ec8d88c