Disable ProfileShortcutManager for App Launcher (chrome.exe --show-app-list)
The existing code in ProfileShort assumes chrome.exe means Chrome Browser.  As a result, when a profile is created (in particular, when App Launcher is first run after install), it attempts to update Chrome shortcut, and ends up creating it, even though Chrome is not installed.

The minimal-effective solution is to disable ProfileShortcutManager if chrom.exe is run for App Launcher (via chrome.exe --show-app-list).

If we ever want to support profile shortcuts for App Launcher, we'll need to undo this change, and also update

profile_shortcut_manager_win.cc: CreateOrUpdateDesktopShortcutsForProfile()

to use ChromeAppHostDistribution, and scan for app_host.exe in shortcuts.

BUG=174849


Review URL: https://chromiumcodereview.appspot.com/12570009

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@188121 4ff67af0-8c30-449e-8e8b-ad334ec8d88c