Fix "advanced" interstitial button for GAIA destinations.
The interstitial "advanced" button/link triggers when the users is redirected off of the various GAIA login pages.  This works when the destination in gmail.com but not when the destination is a GAIA page itself (like for enabling SMS auth).

Now check for reaching the "continue" URL as a valid trigger.

Extracting the "continue" URL needed updates, too:
- examine the most recent URL request, not always the first one (which may have been a POST and thus not have any continue=value parameter)
- stop stripping "continue" embedding when we reach the end (not just upon leaving the GAIA domain)


BUG=179172


Review URL: https://chromiumcodereview.appspot.com/12512004

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@188166 4ff67af0-8c30-449e-8e8b-ad334ec8d88c