Fix a use after free bug in PersistentTabRestoreService
In current code, "if (is_canceled.Run())" is checked before posting the callback to main thread (from FILE thread). The callback is "PersistentTabRestoreService::Delegate::OnGotLastSessionCommands()". So if the delegate is deleted after the callback posted but before callback run, the callback will still run because it does not check is_canceled, hence a use-after-free.

Now RunIfNotCanceled is run on main thread and will check "is_canceled", this prevents running the callback if delegate is deleted.

BUG=180051

Review URL: https://chromiumcodereview.appspot.com/12778004

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@188348 4ff67af0-8c30-449e-8e8b-ad334ec8d88c