Fixed CommandQueue javascript errors when loading an image failes.
It is possible that ImageEditor is loaded with an empty canvas. This may happen if an image is too big, or it contains errors. Then, CommandQueue got initialized with an empty canvas, and then were failing when trying to copy that canvas. This patch fixes this problem by checking if the source canvas is empty before copying.

TEST=Open a huge image, which is too big to handle by Files.app (eg. 20mpix).
BUG=196449


Review URL: https://chromiumcodereview.appspot.com/12861003

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@188412 4ff67af0-8c30-449e-8e8b-ad334ec8d88c