Improve mouse/trackpad handling when closing a context menu
(Re)Implemented event reprocessing when a context menu is
exited via a Mouse Press outside of the menu. The menu is
dismissed, its message loop is exited and then the mouse
event is reprocessed with the expected result like:
activation of another window, a button or bringing up
another context menu.

BUG=188675
TEST=manual; Bring up context menu; press mouse to 
exit on background, another window, or control element
and the "normal" result.


Review URL: https://chromiumcodereview.appspot.com/12342023

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@188424 4ff67af0-8c30-449e-8e8b-ad334ec8d88c