Fixed artifacts in drag-and-drop thumbnail in Files.app.
When trying to drag-and-drop a file which is not draggable (eg. not available offline, while Drive is not available), then the thumbnail had not been cleaned causing artifacts during the next (but successful) drag-and-drop try on a draggable file. Also, when clicking fast, sometimes two thumbnails were displayed. This patch solves this issue by avoiding generation of a thumbnail when dragging event is cancelled.

TEST=Disconnect from Internet. Try to drag a file which if not available offline. Later, try to drag another draggable file. Thumbnail should be fine.
BUG=181508


Review URL: https://chromiumcodereview.appspot.com/12739005

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@188588 4ff67af0-8c30-449e-8e8b-ad334ec8d88c