[mac] App shims
The AppShimHostController uses an IPC::ChannelFactory to listen on a socket for
app shims connecting. App shims send a launch message when they start, so we
translate that into actually launching the app.

I changed chrome_main_app_mode_mac.mm to use the IPC channel instead of sending
an Apple Event as before, and added an event loop and an IO thread.

BUG=138733


Review URL: https://chromiumcodereview.appspot.com/12623005

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@188675 4ff67af0-8c30-449e-8e8b-ad334ec8d88c