Change the way the omnibox handles CTRL-Enter navigations.
Previously, holding CTRL in the omnibox caused the providers to change their top result, promoting a URL_WHAT_YOU_TYPED result with a ".com" suffix.

With this patch, the providers now disregard the CTRL state and, instead, when a CTRL-Enter is encountered the OmniboxEditModel tries to replace the top AutocompleteMatch by a URL_WHAT_YOU_TYPED result with the ".com" suffix.

As a result, CTRL-Enter is now supported in instant-extended.

BUG=172223,170124
TBR=isherman@chromium.org,avi@chromium.org

Review URL: https://chromiumcodereview.appspot.com/12485004

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@188738 4ff67af0-8c30-449e-8e8b-ad334ec8d88c