Fix crash on browser termination and Panel::Close() using nested loop on Mac.
The crash happens when a deferred destruction of the last BrowserWindowController causes a "app is terminating' notification to be sent and panels (which are not browser windows) start to close. We call performClose: on them and this contains a nested message loop. I suspect, although not sure because I could not reproduce the crash, that subsequent reentrancy causes us to use deallocated objects.

This is a speculative attempt to see if deferring of panel close will help...

BUG=180445


Review URL: https://chromiumcodereview.appspot.com/12811004

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@188868 4ff67af0-8c30-449e-8e8b-ad334ec8d88c