Use BookmarkModel as non-const pointer.
We are passing BookmarkModel by non-const reference in model_test_utils::AddNodesFromModelString, but that is not allowed by Google C++ style guide. And in fact we should be just passing a non-const pointer instead. Passing a non-const pointer to BookmarkModel is also the canonical way to pass it through functions.

http://google-styleguide.googlecode.com/svn/trunk/cppguide.xml?showone=Reference_Arguments#Reference_Arguments

BUG=208333
R=sky@chromium.org,asvitkine@chromium.org

Review URL: https://chromiumcodereview.appspot.com/12513012

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@188941 4ff67af0-8c30-449e-8e8b-ad334ec8d88c