drive: Run fast-fetch during delta feed fetching.
Previously, LoadIfNeeded immediately returned if already loaded().
But if it is refreshing() and it is a directory load, we need to
invoke fast-fetch, otherwise a large delta feed may take long time
to update directory list.

Along the way, moved the "LoadIfNeeded" logic from DriveFileSystem
to ChangeListLoader.

BUG=178348

Review URL: https://codereview.chromium.org/12620012

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@189003 4ff67af0-8c30-449e-8e8b-ad334ec8d88c