Bookmarks Bar is displayed when visiting about:crash after chrome://newtab.
Even if a renderer was crashed by chrome://crash (or about:crash), Browser couldn't detect it.

I added code to make Browser handle it when a tab is invalidated.  Besides, I fixed
BookmarkTabHelper::ShouldShowBookmarkBar() so that it returns false when an instance
of WebContents doesn't run.

BUG=30175
TEST=Manual

Review URL: https://chromiumcodereview.appspot.com/12330046

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@189153 4ff67af0-8c30-449e-8e8b-ad334ec8d88c