drive: Pass through AboutResouce from server amond callbacks in ChangeListLoader.
This help us refactoring the pseudo tree structure with separate namespaces.
We would like to create a dummy "My Drive" root in ChangeListProcessor and use
it as "drive/root" instead of having it in DriveResourceMetadata.
This operation needs the root folder ID.
It also helps our migration to Drive API.

BUG=174233
TEST=unittests. Open Files App and verify Google Drive is available.


Review URL: https://chromiumcodereview.appspot.com/12588009

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@189323 4ff67af0-8c30-449e-8e8b-ad334ec8d88c