Suppress push messaging invalidations triggered by extension install.
When a push messaging extension is installed, it gets registered with
the cache invalidation server so it is informed of future push messages.
However, the cache invalidation server (Thialfi) provides historical
context on registration, installing a push messaging extension always
results in immediate invalidations on each of its subchannels.

These invalidations are likely to be confusing and of low value to the
extension. Instead, simply suppress the first invalidation on each
subchannel when an extension is installed.

The test also uses a random payload now, to ensure that it's getting
accurate results rather than an old payload cached from previous
attempts.

BUG=166189,179190

Review URL: https://codereview.chromium.org/12880006

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@189394 4ff67af0-8c30-449e-8e8b-ad334ec8d88c