chromeos: Add idle warning prefs.
This adds new "idle warning" prefs that are pushed to the
power manager to tell it how long to wait before sending
IdleActionImminent signals.  If the inactivity timer is
reset before the idle action is performed, an additional
IdleActionDeferred signal is sent.  This can be used to
determine when to show and dismiss a "You will be logged out
soon" dialog.

BUG=195560


Review URL: https://chromiumcodereview.appspot.com/12475015

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@189397 4ff67af0-8c30-449e-8e8b-ad334ec8d88c