Fixed a long-standing browser action icon order randomization bug.
We've had bug reports about the order of browser action icons in the toolbar getting randomized but we were never able to reproduce it or get actionable reports from users (until now).

The problem is with more than one browser action extension crashing at the same time (which can occur silently). The fix is to remember the order of not just the currently active icons (and the last extension that went away), but icons for all installed extensions. Added an automated test to confirm.

Code removed:
1) last_extension_removed_* as this CL supersedes that.
2) The action box implementation for browser actions, as it was not used anymore and was slated for deletion (would have otherwise had to come up with a fix for that as well).
3) Debugging traces from a test that were meant to give insight into the test timing out but it no longer seems to be.

TBR=sky
BUG=68414
TEST=Browser actions should behave as before, except if you shoot down multiple extensions with the Task Manager and re-enable them in random order they should have the same order as they had before they crashed. Also, the order of an extension should be forgotten once the extension is uninstalled and it should reappear last when reinstalled.
Review URL: https://codereview.chromium.org/12622011

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@189572 4ff67af0-8c30-449e-8e8b-ad334ec8d88c