Fixing problem with a tabless browser windows showing up after a task manager was used
I traced it down to the creation of the task manager which grabs an existing (or new) browser. It requires the browser only for two things: determining the system it is running in and for giving the created web dialog a context (a window which is used (by Aura only) to position the window within the hierarchy.

As I can see it there are no other dependencies to the browser and it can be disposed of after the task manager was created (and there are no tabs in the tabbed browser).

BUG=176670
TEST=visual


Review URL: https://chromiumcodereview.appspot.com/12712018

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@189579 4ff67af0-8c30-449e-8e8b-ad334ec8d88c