Keep the 'fetching files' message until the Drive root is ready to be read.
For introducing the 'fast-fetch' feature, getDirectory(root)
is changed not to trigger feed loading anymore. To wait for
Drive really ready for operations, we need to wait for directoryReader.readEntries() to finish.

BUG=197246
TEST=Open Files.app > "Google Drive", ctrl-click gear menu and clean local cache.
Verify spinner with some text "fetching files" is shown, not just a spinner animation.

Review URL: https://codereview.chromium.org/12978002

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@189599 4ff67af0-8c30-449e-8e8b-ad334ec8d88c