chromeos: Remove dead power manager code.
This removes the OutputObserver class, which used to listen
for ScreenPowerSetSignal D-Bus signals from powerd and pass
the requests on to OutputConfigurator.  This is now handled
by DisplayPowerServiceProvider, which serves as an endpoint
for SetDisplayPower method calls.

It also removes the timestamp argument from
HandleUserActivity method calls to powerd.  powerd no longer
uses (or reads, for that matter) the argument.

BUG=chromium:221391
TBR=ben@chromium.org

Review URL: https://codereview.chromium.org/12992004

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@189809 4ff67af0-8c30-449e-8e8b-ad334ec8d88c