Generalzie GpuBlacklist to GpuControlList.
This CL refactor the GpuBlacklist so it can be configured to handle a different set of features.

So it can handle not only blacklist, but dual gpu switching, and driver bugs.

I will add the code to truly manage driver bug workarounds in the next CL.

BUG=222857
TEST=content_unittests
TBR=asargent
Review URL: https://codereview.chromium.org/12976004

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@189889 4ff67af0-8c30-449e-8e8b-ad334ec8d88c