Add RenderWidgetHost::GetSnapshotFromRenderer method to content/ interface for snapshotting for cases where we might not be able to use CopyFromBackingStore because the browser side can't access the backing store or accelerated surface due to driver issues or WinXP.
Consolidate renderer-side snapshotting into that method and move clients (tabsApi, tabCapture and NTP) to the new method.

Let the Linux CopyFromCompositingSurface always use this since the current one is incorrect if the tab is in the background or is covered by a window.

Remove Linux-GTK snapshotting workarounds.

BUG=188867, 174957, 132301

Review URL: https://codereview.chromium.org/12881005

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@189969 4ff67af0-8c30-449e-8e8b-ad334ec8d88c