Preventing race conditions in Google Now extension.
Introducing a task manager. A task is an event chain that does certain work that we want to be atomic. Execution of multiple tasks is serialized.

Also, adding checks for anomalous situations, such as a forever-running task keeps the event page from unloading or a task still running when the event page unloads.

BUG=164227

Review URL: https://chromiumcodereview.appspot.com/12316075

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@190023 4ff67af0-8c30-449e-8e8b-ad334ec8d88c