Replace screen capture confirmation infobar with a message box.
Currently an infobar is used to get user consent for screen capture API. We
are moving away from infobars for that API. The plan is to use modal prompt 
instead of infobars (but eventually they may be replaced with something 
better). This CL adds ScreenCaptureConfirmationUI interface, so that it will
be easy to replace infobars. The new ScreenCaptureConfirmationUIInfobar 
implements this interface using infobars (will be removed eventually).

BUG=196287
TBR=thakis@chromium.org

Review URL: https://chromiumcodereview.appspot.com/12843009

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@190149 4ff67af0-8c30-449e-8e8b-ad334ec8d88c