Alternate NTP Mac: Show overlay at fixed position
Previously the overlay was shown at the bottom of the toolbar. The height of
the bookmark bar changes if the bookmark bar is attached. This meant that the
overlay would change positions (relative to the omnibox) depending on the
bookmark bar's attached state.

With this CL we now show the overlay at the fixed offset from the omnibox.

One side effect of this is that the bottom of the toolbar can no longer act
as a separator for the overlay. I've changed the overlay to draw its own
separator.

Screenshot: http://i.imgur.com/ub9gfDo.png

BUG=173475, 174077
TBR=sky@chromium.org


Review URL: https://chromiumcodereview.appspot.com/12674019

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@190213 4ff67af0-8c30-449e-8e8b-ad334ec8d88c