drive: Add a hook to update the default download directory.
Update the user's default download directory setting if necessary.
We are going to change "/special/drive" to "/special/drive/root" in order to support more Google Drive features
and need to update the user's setting if it points to Drive.
This CL adds
- a utility function to check if migration is necessary or not.
- a utility function to insert the namespace "root" in FilePath
and uses them in preferences.cc to update the pref.

BUG=174233,180499
TESTS=unittests


Review URL: https://chromiumcodereview.appspot.com/12706023

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@190356 4ff67af0-8c30-449e-8e8b-ad334ec8d88c