Action choice dialog should not auto-choose a remembered item when run with shift modifier pressed.
The action choice dialog remembers the choice, so when an sd/usb with dcim folder is inserted again, then the remembered app is automatically started. This patch adds an option to avoid this auto-launching, and instead always showing the action choice dialog, when shift key is pressed. This is for power users.

TEST=Enter sd/usb with dcim and choice Import to Drive. Hit 'Always'. Do this again with shift, and later without shift.
BUG=222941


Review URL: https://chromiumcodereview.appspot.com/12880012

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@190625 4ff67af0-8c30-449e-8e8b-ad334ec8d88c