Omnibox: Revise Decision that Default Match Has Changed
This changes the logic that decides whether the default match has changed
to look also at the keyword.  This should fix both of the linked bugs
(a switch of match type such as from a URL to a search will be a switch from
no keyword to a keyword; a switch of a template URL should be a change
in keywords).  Both of the bugs listed are purely theoretical--they lack
repro steps.  Hence, I'm pretty sure this is right but I have no way to
test it.  If you have suggestions for testing and would like me to do it,
tell me.

BUG=103468,159736


Review URL: https://chromiumcodereview.appspot.com/13001005

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@190636 4ff67af0-8c30-449e-8e8b-ad334ec8d88c