Fix the no password save issue for ajax login
The idea is that if a navigation is triggered by non-user gesture after a password form is submitted, then we assume that this navigation is related to the previous form submit and chrome should prompt for saving password if all criteria for provisionally saving password are met. 

The link between the previous form submit and the coming navigation is weak, so its possible for chrome to inherit password forms from previous state when it should not. However most false cases would be aborted at PasswordManager::ProvisionallySavePassword, which does a list of sanity checks for saving password, such as if the form is already seen on the current page. 

BUG=43219


Review URL: https://chromiumcodereview.appspot.com/12713007

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@191024 4ff67af0-8c30-449e-8e8b-ad334ec8d88c