Fix a bug wherein a member method masked a windows API call.
ProcessSingleton is supposed to flash/bring to the foreground the visible modal dialog in response to a new execution of chrome.exe. I verified that the functionality was broken by calling chrome.exe --try-chrome-again=0, sending the toast to the background, and then launching chrome.exe again. Before this CL, nothing happens. After this CL, the toast pops to the foreground.

R=thakis@chromium.org
BUG=None


Review URL: https://chromiumcodereview.appspot.com/12918031

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@191134 4ff67af0-8c30-449e-8e8b-ad334ec8d88c