chromeos: Convert google_apis::ResourceEntry to DriveEntryProto sooner to reduce memory usage
ResourceEntry is much larger than DriveEntryProto since it contains almost all information from the server, including what is not necessary to create DriveEntryProto.
This change makes the ResourceEntry-to-DriveEntryProto conversion as soon as ResourceEntry is received from the server, to avoid filling the memory up with all ResourceEntry in the user's account.

With this change, we can reduce the memory consumption with change list loading by about 60% (280MB to 100MB with 45000 files)

BUG=223860
TEST=unit_tests


Review URL: https://chromiumcodereview.appspot.com/13116002

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@191154 4ff67af0-8c30-449e-8e8b-ad334ec8d88c