Split FaviconHelper in two: ImageLoadingHelper and FaviconHelper.
This is the first step of fixing http://crbug.com/196769.

The problem is that there are two types of users of Content::DownloadFavicon() - the favicon consumers and regular icons and images consumers. Recently, the otherwise-generic image download+decoding implementation regressed because it was assumed to be only used for favicon loading. 

The proposed fix is to add a parameter to the method (enum, FAVICON/IMAGE) and rename it from DownloadFavicon to DownloadImage, to make sure the name corresponds to the impl/usage.

This is rename-only part, no additional parameter yet. The FaviconHelper class was split in 2, with new ImageLoadingHelper dealing with images. 


BUG=196769


Review URL: https://chromiumcodereview.appspot.com/12780024

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@191304 4ff67af0-8c30-449e-8e8b-ad334ec8d88c