Fixed copying in Files.app v2.
This patch addresses two issues around copying files in Files.app working as a platform app.

Firstly, in Files.app v2, we can't close the background page while there are any Files.app's windows active, since the background page
stores handles to these windows. If we prematurely close the page, then we loose that information, which breaks communication between the background page and windows. This patch fixes this issue by not letting the FileCopyManager close the background page after 5 seconds, if there are any windows visible.

Secondly, passing a root of the local filesystem from a window to the background page (FileCopyManager) is unreliable. It works, but after a window is closed, completion callbacks are not called. This patch solves this issue by requesting the filesystem in the background page, instead of passing it from the window.

TEST=Run Files.app with --file-manager-packaged. (1) Copy a file, wait 30 seconds, copy again. Butter bar should show in each case. (2) Copy a big file, close the window while copying, open Files.app again, and wait until the copying is finished.
BUG=224214

Review URL: https://chromiumcodereview.appspot.com/13301002

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@191357 4ff67af0-8c30-449e-8e8b-ad334ec8d88c