Propegate setting autohide behaviour to prefs
When the ShelfLayoutManager changes the autohide behaviour this needs to be
propegated to the ChromeLauncherController otherwise when other parts of the
system query the autohide status they can end up with an incorrect value. This
resolves most of the outstanding issues with the "Autohide Launcher" checkbox in
the context menu.

These is still one issue, specifically if you set autohide off, swipe the
launcher off the screen and then make it reappear, via either swiping or mouse
over. The checkbox will not be checked in this case, but the moment you touch or
click off of the launcher it will hide. After hiding the checkbox will be set as
expected, thus partially resolving that fact that you had to turn on autohide
before you could turn it off. This issue due to the fact that the
ShelfLayoutManager is actually turning off autohide, but setting a handler to
watch for future events and when you interact outside of the launcher setting
autohide on again. Resolving this is outside the scope of this CL/bug since it
either requires changing our UI behjaviour or plumbing information about being
in this state to the LaunchContextMenu, both of which would be unlikely to be
backported to 27.

BUG=chromium:173295
TEST=Visually confirmed that the checkbox is repersenative of the state of
     autohide or is only transistorially incorrect.

Review URL: https://chromiumcodereview.appspot.com/12636012

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@191376 4ff67af0-8c30-449e-8e8b-ad334ec8d88c