Introduce browser-agnostic web contents modal dialog positioning abstraction and implement for Views
Support positioning of the web contents modal dialog on a host window,
independent of the type of the host window. Also, support repositioning
the window when the host window geometry changes.

Implement the abstraction for the Views browser window and Views
web contents modal dialog.  The implementation moves the dialog to
the top overlapping the browser chrome and is enabled behind the
--enable-new-dialog-style flag.

Before/after screenshots at
https://code.google.com/p/chromium/issues/detail?id=157161#c46.  Visual
artifacts/incorrect layout within the client area are a result of the new style
dialog being a work in progress and will be addressed separately.

BUG=157161

Review URL: https://chromiumcodereview.appspot.com/13212007

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@191735 4ff67af0-8c30-449e-8e8b-ad334ec8d88c