Don't generate protobuf fields nor decoding code for unsupported policies.
The cloud policy protobuf is generated by a python script from the policy
templates file; a helper function that converts those protobufs to base::Values
is also generated. This change skips the protobuf and code generation for
policies that are not supported on the target platform.

It also filters policies that have a limited 'supported_on' version for the
matching platform; currently this only affects the HideWebStorePromo policy.

Changes in a linux chromeos=1 Release build:
- CloudPolicySettings goes from 216 to 172 fields
- cloud_policy_generated.cc goes from 4337 to 3193 lines
- chrome binary size decreases 45KB, of a total of 48MB

BUG=141368


Review URL: https://chromiumcodereview.appspot.com/12987009

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@191832 4ff67af0-8c30-449e-8e8b-ad334ec8d88c