Skip loading installed extensions when --install-from-webstore is active.
Loading installed extensions was interfering in a bad way with the --install-from-webstore abbreviated workflow. In particular, if an extension subscribed to the onStartup event and showed some UI when it came:
1) The UI was disabled, while the modal install dialog was active.
2) DCHECK in ProfileDestroyer would fire when Install or Cancel was clicked in the dialog.
3) As a result, the new extension wouldn't get installed.

Since loading installed extensions makes little sense for the --install-from-webstore workflow, there is no harm in just not loading them.

BUG=164183


Review URL: https://chromiumcodereview.appspot.com/12858005

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@191964 4ff67af0-8c30-449e-8e8b-ad334ec8d88c