Non-web-accessible extension URLs should not load in non-extension processes
This is a slightly modified version of my previous CL: https://codereview.chromium.org/12218064/.
The only difference is that we allow any resource request to succeed, if the extension has any web_acessible_resources. The reason for that we have been lax and allowed subresource loads, even if they are not explicitly added to the manifest (see crbug.com/179127 for details). This should be tightened up with a v3 manifest requirement to explicitly list all subresources.

BUG=173688


Review URL: https://chromiumcodereview.appspot.com/12457042

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@192121 4ff67af0-8c30-449e-8e8b-ad334ec8d88c