Revert issue 158379
This reverts crbug.com/158379 which changed the apps v1 window creation behavior from outer bounds to inner bounds. That change broke "expected" behavior for a class of extensions, and made the behavior inconsistent between Create and Get/Update. There was some internal discussion about what expected behavior should be, and we decided to revert the behavior.

A test was added for consistency between Create and Window properties, and the documentation was clarified.

BUG=173831

Review URL: https://codereview.chromium.org/13375002

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@192181 4ff67af0-8c30-449e-8e8b-ad334ec8d88c