Added a PolicyCertVerifier that uses the trust anchors from the ONC policies.
The MultiThreadedCertVerifier can optionally use a CertTrustAnchorProvider to
get a list of additional certificates to trust, without importing them into the
NSS database. This CL wraps the MultiThreadedCertVerifier with a custom verifier
that includes a trust anchor provider.

The trust anchor provider returns all the certificates from the user ONC policy
that have the Web trust flag. The PolicyCertVerifier also writes a preference
in the Profile once any such certificate is used.

This feature is currently behind a flag, until a warning UI is implemented.
The warning should be displayed if UsedPolicyCertificates() is true for the
given profile.

TBR=pneubeck@chromium.org
BUG=216495

NOTE: this is a reland of https://codereview.chromium.org/13035003, with tests fixed.


Review URL: https://chromiumcodereview.appspot.com/13532005

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@192324 4ff67af0-8c30-449e-8e8b-ad334ec8d88c