Use last_selected_directory() when invalid path is passed to CrOS file picker.
This is for working around the case when <input type="file"/> control is used
twice; CrOS file picker kindly resolves a remote file path to a local cache
and the type=file control tries to default to the path, but the cache path
itself is out of user-visible directory. In such a case, we just use the
last_selected_directory(), rather than going back to the very default directory.

BUG=178013

Review URL: https://chromiumcodereview.appspot.com/13614002

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@192492 4ff67af0-8c30-449e-8e8b-ad334ec8d88c