Fix launcher and system tray accessibility.
Fix AccessiblePaneView's logic when "Escape" is pressed: now it
restores focus to the last focused view but only if it's within the same
window and not within this pane - in practice, that only happens within
the browser window. Otherwise, it deactivates the current window - so
pressing Escape while focused on the launcher or status area will restore
activation to the browser window.

Also adds an accessible notification when the system tray opens.

BUG=225920


Review URL: https://chromiumcodereview.appspot.com/13471014

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@192734 4ff67af0-8c30-449e-8e8b-ad334ec8d88c