Cancel notification toast timeoout for webkit ones.
Currently webkit notifications are indistinguishable from TYPE_SIMPLE.
Instead of adding a new type, it just adds a flag to cancel the timeout
and only the webkit notification code path sets it.

Note that this 'no-timeout' flag is slightly different from the persistency
of the system notification for ChromeOS. In addition to no-timeout, the
system notification should not disappear for open/close of message-center,
i.e. MarkShownAsPopup should be canceled too. It will need an additional
flag.

BUG=229014


Review URL: https://chromiumcodereview.appspot.com/13877002

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@193315 4ff67af0-8c30-449e-8e8b-ad334ec8d88c