Fixed cache eviction in Files.app's slide mode.
Sometimes, prefetching was fast, which caused evicting the image being loaded from cache. As a result, full resolution image was not being loaded properly. This patch addresses this issue by evicting the full-resolution cache in advance, before loading an image from cache. As a result, there is always a space for cache, so there is a guarantee that the image just fetched from cache will not be evicted by the image being prefetched.

TEST=Follow the bug's description.
BUG=227264


Review URL: https://chromiumcodereview.appspot.com/13775010

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@193324 4ff67af0-8c30-449e-8e8b-ad334ec8d88c