Remove root resource id aliasing from DriveResourceMetadata.
This is a part of Drive API v2 migration. On Drive API v2, the root resource
id is not fixed, while it is fixed on GData WAPI, which is "folder:root".
By this CL, the MyDrive root entry will be created when the "drive" (grand
root) is accessed for the first time, instead of creating it in
DriveResourceMetadata::Initialize.
Along with the change, this CL also implements the Fast-Fetch behavior
in ChangeListLoader::DoLoadDirectoryFromServcer.

BUG=174237
TEST=Ran unit_tests and tested manually on chromeos-chrome


Review URL: https://chromiumcodereview.appspot.com/13866009

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@193362 4ff67af0-8c30-449e-8e8b-ad334ec8d88c