Consider a RenderView dead as soon as the RenderViewHost starts to delete it
The RenderViewHost will start to filter out IPCs related to provisional loads,
however, other load related IPCs, namely those handled by the WebContentsImpl,
get still forwarded.

This confuses the webNavigation API when it sees a URL finishing to load that it
never saw the provisional load for.

This led to a bug in SessionRestoreTest.RestoreWebUISettings: after the test
finished, InProcessBrowserTests::QuitBrowsers tears down all browsers. If the
RenderView did not yet commit its provisional load at that time, but manages
to commit and finish the navigation before processing the close IPC a DCHECK
in the webNavigation API was triggered.

BUG=176304
TEST=SessionRestoreTest.RestoreWebUISettings is no longer flaky

Review URL: https://codereview.chromium.org/13845007

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@193631 4ff67af0-8c30-449e-8e8b-ad334ec8d88c