In the bookmark manager, right clicking on an empty in the bookmark list should allow all displayed bookmarks to be opened.
In the scenario where the user has initiated the context menu in the
bookmark list without any bookmarks selection the target bookmarks
for open commands will be the entire contents of the Bookmark List
(the folder selected).

The change involves modifying the |canExecute| and the |handleCommand|
so that the entire list is considered in the above scenario.

Note that the menu items that are actionable are only the 'Open all...'
commands and not 'Cut', 'Copy', and 'Delete'.

BUG=225005
TEST=Follow reproductions steps; right click on a blank area of the
Bookmark List to see the now actionable 'Open all..." command and that
they indeed launch the proper bookmarks.

There is a (seemingly) unrelated feature that uses the modified code
and that is double clicking on a bookmark or selecting bookmark(s)
and pressing enter.  These actions should remain unchanged.

Review URL: https://chromiumcodereview.appspot.com/13373002

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@193704 4ff67af0-8c30-449e-8e8b-ad334ec8d88c