History: Add name of the originating device to the visit dropdown menu.
For visits from other devices, show the device name at the top of the
dropdown menu for that visit. For local visits, nothing is shown.

Additionally, this CL removes some dead code and fixes a small bug
where visits could be incorrectly grouped based on the server's
timezone, not the local device's timezone.

BUG=180968
TEST=With full history sync enabled, open chrome://history and click
on the dropdown menu of a visit from another device. The device name
should be shown at the top of the menu.

Review URL: https://codereview.chromium.org/12729024

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@193942 4ff67af0-8c30-449e-8e8b-ad334ec8d88c