Do not hard-code the avatar icon for a new enterprise profile.
Updated SigninManager to call the appropriate ProfileManager API to generate
a locally-unique avatar icon when creating a new profile during enterprise
signin. Sync can still clobber this unique icon with whatever the setting was
on the most-recently-synced device (which may not be unique) but this is an
improvement over the old hard-coded behavior.

BUG=179831

Review URL: https://chromiumcodereview.appspot.com/14234004

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@193981 4ff67af0-8c30-449e-8e8b-ad334ec8d88c