Fix opening Google Doc's documents in Files.app V2.
Opening such files failed, when no browser window was running. This is because packaged apps are not running within a browser window, and therefore GetCurrentBrowser returns NULL, if there is no browser window running. Therefore we need to create a browser instance. This patch fixes this problem by replacing GetCurrentBrowser with a FindOrCreateTabbedBrowser call.

TEST=Run chrome with --file-manager-packaged, launch Files.app and try to open a .gdoc file.
BUG=231190

Review URL: https://codereview.chromium.org/13980006

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@194287 4ff67af0-8c30-449e-8e8b-ad334ec8d88c