Speculative fix for instant overlay crash
We're getting some crashes in
-[OverlayableContentsController setOverlay:...].

I wasn't able to reproduce the crash. My guess is that we're somehow
referencing a stale WebContents pointer.

This is a speculative fix that clears the WebContents reference when
the object is destroyed. I'll revert this change if we continue to
see crashes.

BUG=230880

Review URL: https://codereview.chromium.org/14242004

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@194440 4ff67af0-8c30-449e-8e8b-ad334ec8d88c