Complete (but inefficient) implementation of the image retargetting method.
This CL contains:
1. Convolution of arbitrary channel with arbitrary kernel (convolver*).
2. Gaussian gradient magnitude implementation.
3. Image profile (X and Y projections) computations.
4. Otsu-like thresholding of profiles.
5. Image decimation.
6. The main routine which binds it all together.

Note: 1 and 2 do work, but remain main sources of suckiness due to performance
problems. I actively work on this. Still, I'd love to get the current state in
to establish a baseline and for viewing pleasure of those who are interested.

BUG=155269

Review URL: https://codereview.chromium.org/13947013

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@194565 4ff67af0-8c30-449e-8e8b-ad334ec8d88c