History: Workaround bug with clear button in search field.
Due to a Blink bug (http://crbug.com/231830), the clear button in the
search field is not properly rendered if the <header> element has
its transform style property set to anything other than 'none'.

A simple workaround to make it render properly in 99% of cases is to
set the value of -webkit-transform to 'none' instead of 0px.

BUG=231291
TEST=Open chrome://history, type something in the search field, and
ensure that the clear button (a circle with a "X") is visible.

Review URL: https://codereview.chromium.org/14053006

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@194586 4ff67af0-8c30-449e-8e8b-ad334ec8d88c