sync: Finish the SyncScheduler refactor
This change removes SyncSessionJob entirely.  The first step towards
this goal was to refactor all functions that depended on SyncSessionJob.
All these functions have either been inlined or modified to take
different parameters instead.  DoSyncSessionJob was split into
two separate functions, one for configure jobs and one for nudge jobs,
which removes the need for SyncSessionJob's "purpose" member.

The SyncScheduler's pending_configure_job_ has been replaced with
a ConfigParams member, since that was the only part of the job still
being referenced.  The pending_nudge_job_ has been replaced with
scheduled_nudge_time_ (which is similar to the old scheduled_start_
member of SyncSessionJob) and a new object called a NudgeTracker.

The NudgeTracker inherits the SyncSessionJob's responsibilities with
respect to coalescing nudge sources.  The plan is to extend this class
to support more sophisticated nudge coalescing in the future.  For now
it tries to emulate the old SyncSessionJob behaviour as closely as
possible.

Some of the refactoring does change behaviour.  In particular, the
decision-making logic has been updated to fix issues 179515 and 155296.

BUG=175024,179515,155296

Review URL: https://chromiumcodereview.appspot.com/13743003

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@194766 4ff67af0-8c30-449e-8e8b-ad334ec8d88c