Add a switch to Chromium to enable/disable 'image/webp' accept header.
This switch is off by default. It will be turned on later once we verify that this change doesn't break any sites.

It can be can be enabled through command-line (--enable-webp-in-accept-header) or through 'chrome://flags'.

Note: This is a follow up of the pre-requisite change in Blink: https://codereview.chromium.org/13814024/

BUG=169182

Review URL: https://chromiumcodereview.appspot.com/14273007

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@194853 4ff67af0-8c30-449e-8e8b-ad334ec8d88c