[Mac] Fix handling of download shelf auto close.
If the mouse is not on the download shelf when an auto-close is
scheduled, the shelf would not close automatically. Fix this by
closing immediately if the mouse isn't on the shelf.

Also adds unit-tests and update the shelf to use CrTrackingArea
instead of NSTrackingArea.

BUG=none
TEST=(1) Start a download that opens automatically.
     (2) Move the mouse off the shelf and wait for completion.
     (3) The shelf should close automatically.

Review URL: https://chromiumcodereview.appspot.com/13318002

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@195080 4ff67af0-8c30-449e-8e8b-ad334ec8d88c