History: Reload page only after web history deletion has completed.
When history sync is enabled, deletions need to be done in both the
history DB and on the history server. Wait until the request to the
server has completed before reloading the page, otherwise it can
appear that the deletion did not take effect.

BUG=None

Review URL: https://codereview.chromium.org/14225004

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@195140 4ff67af0-8c30-449e-8e8b-ad334ec8d88c