Disable history delete directives when sessions are encrypted.
Full history sync is based on server-side process of session data, so
if session data is encrypted, it can't be read on the server. In this
case, delete directives are also unnecessary, so they can be disabled
whenever the session data type is encrypted.

BUG=229015

Review URL: https://codereview.chromium.org/14139010

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@195154 4ff67af0-8c30-449e-8e8b-ad334ec8d88c