Remove --show-app-list-shortcut flag and implement new app launcher enable logic.
The app launcher now adds itself to the start menu on the first run of chrome
past this patch.

The webstore enable of the app launcher will create shortcuts on the desktop
and pin an icon to the taskbar.

The app launcher should never be disabled after it is enabled.

BUG=233434
TBR=benwells@chromium.org

Review URL: https://chromiumcodereview.appspot.com/13940006

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@195158 4ff67af0-8c30-449e-8e8b-ad334ec8d88c