Update message center systray icon.
New design shows a notification bubble on a gray background when there
are no unread notifications.  When unread notifications exist, an image
displaying the number of notifications (up to 9+) is rendered on a blue
background.

The new png is a sprite map of the digits.

TEST=manually generate 11 notifications
BUG=189129

Review URL: https://codereview.chromium.org/14192025

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@195314 4ff67af0-8c30-449e-8e8b-ad334ec8d88c