Fixed a performance regression after introducing the new ui flag.
In Files.app, it is often checked if the Files.app is in the new UI mode. To do that, it checked for the app's version in the manifest. This operation is expensive and caused a performance regression. This patch addresses this issue by caching the value.

TBR=yoshiki@chromium.org
TEST=Tested manually. Scrolling should be fast with lots of items.
BUG=226912

Review URL: https://codereview.chromium.org/13943006

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@195459 4ff67af0-8c30-449e-8e8b-ad334ec8d88c