Refactoring: replace SearchInDirectory by SearchByTitle.
The requirement of SearchInDirectory is just title exact-match search.
This CL is the preparation of it by replacing the method by SearchByTitle. Also
this moves the responsibility of escaping (and query formatting) from sync
drive service to google_apis. This is because the query formatting is different
between GData WAPI and Drive API v2, and the gap is filled in
DriveServiceInterface layer. At the moment, this CL shouldn't change the
behavior yet (i.e. still partial match).

BUG=232352
TEST=Ran unit_tests

Review URL: https://chromiumcodereview.appspot.com/14146006

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@195528 4ff67af0-8c30-449e-8e8b-ad334ec8d88c