[Android] Refactor NativeView to be able to use it for AutofillDialog.
Create a new class WindowAndroid that serves the purpose of representing a native (platform specific) view where Chromium code expects to have a cross platform handle to the system view type. As Views are Java objects on Android, ViewAndroid.java and view_android.* will provide the expected abstractions on the C++ side and allow it to be flexibly glued to an actual Android Java View at runtime. It should only be used where access to Android Views is needed from the C++ code as there are easier/better ways to access Views from Java.

This new abstraction will allow to reuse AutofillPopup for regular use in a webpage (using ContentView) as well as for AutofillDialog.

BUG=229199

Review URL: https://chromiumcodereview.appspot.com/14018004

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@195750 4ff67af0-8c30-449e-8e8b-ad334ec8d88c