Close syncSetup overlay if the user isn't signed in yet.
If the user tries to navigate to http://settings/syncSetup (either directly
or by reloading the page while in the middle of sync setup) we will redirect
the user back to http://chrome/settings if they are not signed in yet and
web-based signin is enabled.

The old code to support initiating signins through the settings page is obsolete
because web-based signins don't need to route through settings any more.

BUG=229836

Review URL: https://chromiumcodereview.appspot.com/14293007

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@196115 4ff67af0-8c30-449e-8e8b-ad334ec8d88c