Supports MessageCenterObserver interface in MessageCenterView (2nd).
Now it receives the add/remove/update events by itself. This is a
preparation CL for repositioning on the message center, but this itself
also allows partial-update in the center.

The previous CL was reverted due to a test crash on linux_chromeos gcc.
This crash doesn't happen on clang, it depends on the order of releasing
members of a class in dtor, but sometimes the message_center_ may be
released before message_center_view. This never happens on the real
world, so added TearDown to ensure releasing message_center_view prior
to message_center.

BUG=224089
TEST=no break on existing tests on gcc build
TBR=dharcourt@chromium.org,dewittj@chromium.org,stevenjb@chromium.org

Review URL: https://chromiumcodereview.appspot.com/14472010

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@196205 4ff67af0-8c30-449e-8e8b-ad334ec8d88c