[rAc Android] Do not dismiss the dialog automatically.
Fot the steady state: on Submit/Cancel we receive a Hide() call from the controller anyway, so there is no need to rely on any "automatic" behavior.
For the editing state: we should never dismiss the dialog.

Note that the Back button is ignored with this CL (it has quite surprising behavior in the editing mode anyway), see http://crbug.com/234477.

BUG=233809
NOTRY=True

Review URL: https://chromiumcodereview.appspot.com/14159021

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@196222 4ff67af0-8c30-449e-8e8b-ad334ec8d88c