Implement Pepper GetDeviceID for Windows
The device ID is generated by mixing a salt which is specific to the users profile and with machine-specific information. This CL also connects the action of clearing the salt with the "Clear browsing data > Deauthrorize content licenses" action.

The ChromeOS implementation of GetDeviceID is unified with the Windows implementation and the old method of generating the IDs has been deprecated.

BUG=235193

Review URL: https://chromiumcodereview.appspot.com/12701002

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@196522 4ff67af0-8c30-449e-8e8b-ad334ec8d88c