History: Use version_info token to ensure consistency after deleting entries.
When visits are deleted from the history server, save the version_info token
from the server response, and pass it along with any future lookup or delete
requests. This ensures that future requests see a version of the data that
is consistent with any deletions that were issued.

BUG=232374

Review URL: https://chromiumcodereview.appspot.com/14031010

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@196694 4ff67af0-8c30-449e-8e8b-ad334ec8d88c