Disable FCM on Mac OS 10.7 when kSkipGpuDataLoading is set
This is effectively the same thing that is done above this diff, to
disable the trial on Windows XP.

The trybots on Mac OS 10.7 use the Apple software renderer. This
particular combination is very hang-prone. Browser and content
shell tests specify kSkipGpuDataLoading, which effectively disables
the GPU blacklist, allowing this configuration to slip through and
hang.

BUG=230931

Review URL: https://chromiumcodereview.appspot.com/14283010

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@196885 4ff67af0-8c30-449e-8e8b-ad334ec8d88c