Fetch extension icon for extension/app crashed notifications.
Since the switch to Rich (non-HTML) Notifications, this notification
has not been displayed with an icon because there is no longer a way
to render arbitrary URLs in notifications.  This adds an ImageLoader
to grab the extension's icon from disk and passes a gfx::Image directly
to the notification.

BUG=233207

Review URL: https://chromiumcodereview.appspot.com/13890004

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@196898 4ff67af0-8c30-449e-8e8b-ad334ec8d88c