Turn off Synced Notifications by default
This will provide us more control over when we turn the feature on for which
branch going forwards.

Also provide a three state flag in chrome://flags to enable and disable synced
notifications, and command line flags to override the default for both enabling
and disabling.

BUG=236045
TBR=dimich@chromium.org

Review URL: https://codereview.chromium.org/14520019

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@196996 4ff67af0-8c30-449e-8e8b-ad334ec8d88c