Async loading of CrOS enterprise enrollment screen assets.
This changelist adds a ResourceLoader that registers a bundle of assets
(HTML, CSS, JS) to be loaded on demand and added to the current DOM.

Currently the assets are loaded as soon as the UI is loaded (see the 
handler in login_common.js), so it's not exactly lazy-loading.  Once 
everyone is convinced this works, the loading can be deferred until the
enrollment screen is actually needed.

BUG=232048

Review URL: https://chromiumcodereview.appspot.com/14373012

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@197394 4ff67af0-8c30-449e-8e8b-ad334ec8d88c