Adds back focus restoration hack that Mike recently nuked. The hack is needed for exactly the case that is documented there. I've added a test case for it to make sure we don't regress it again. Since test is focus based I'm putting the test code in chrome, even though it only needs views types.
To get Mikes tests to work I also made it so that changing the focused view also sets the stored focus view. This way if the window is hidden at the time the focus request we'll restore the focused view correctly when the window is shown.

Lastly I disabled a couple of tests that rely on focus and aren't in interactive_ui_tests.

BUG=125976, 131660, 224591
TEST=covered by tests, in chrome can verify by way of selecting text
on page, minimizing chrome, restoring chrome (click on taskbar) and
make sure page is still focused. Windows specific.
R=ben@chomium.org, msw@chromium.org

Review URL: https://chromiumcodereview.appspot.com/14520013

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@197623 4ff67af0-8c30-449e-8e8b-ad334ec8d88c