Extend touch hit target for autohidden shelf to allow dragging
This moves the edge gesture functionality from the BorderGestureHandler
to the ShelfLayoutManager, so that touches on the edge of the screen
when there is a touch enabled web page will be handled correctly.
Specifically when autohidden the shelf will steal a small portion of
the touch screen so that users may perform a drag operation to make it
visible.
    
This CL also changes the method SetHitTestBoundsOverrideOuter to take an
inset for both mouse and touch instead of having a mouse inset and touch
scale. This change is propagated to all the instances that this method is
used and updates the unit tests.
    
BUG=chromium:223666
TEST=Built and confirmed that dragging in/out works with shelf on
left/bottom/right of screen with page open that handles touch events.
Built and ran modified unit tests.

Review URL: https://chromiumcodereview.appspot.com/14439006

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@197625 4ff67af0-8c30-449e-8e8b-ad334ec8d88c