Better overflow in Type Info section of about:sync
When a per-datatype unrecoverable error occurs, the text associated with
it is displayed in the "Group Type" column of about:sync.  This
description can contain long strings of text without whitespace, and can
cause the table to overflow its div.

This change adds an overflow-x:auto property to the containing div so it
will display a scrollbar if the table outgrows the space allocated to
it.

BUG=236694

Review URL: https://chromiumcodereview.appspot.com/14721003

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@197669 4ff67af0-8c30-449e-8e8b-ad334ec8d88c