Re-enable SinkBMs10LevelDown sync integration test
This test was disabled because it was flaky.  It was flaky because the
PutPredecessor() call at the end of a commit cycle could end up
reordering previously committed bookmarks.

Following r191767, we no longer need to call PutPredecessor() after a
successful commit.  The flakiness should no longer exist.

This change was tested locally with:
sync_integration_tests \
   --gtest_filter='*SinkBMs10LevelDown*' \
   --gtest_repeat=20

BUG=107744

Review URL: https://chromiumcodereview.appspot.com/14696006

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@197675 4ff67af0-8c30-449e-8e8b-ad334ec8d88c