Periodically request a list of spellcheck markers in all renderers for feedback
This CL adds a class SpellingServiceFeedback that will be used for sending
feedback to the spelling service. There's one instance of this class per
profile. SpellcheckService owns the SpellingServiceFeedback instance.
SpellingServiceFeedback periodically requests a list of all document markers in
the renderers. It will use the list to determine the kinds of feedback to send
to the spelling service.

BUG=170514

Review URL: https://chromiumcodereview.appspot.com/14126015

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@197779 4ff67af0-8c30-449e-8e8b-ad334ec8d88c