InstantExtended: improve blacklisting and retry logic.
This change gets rid of blacklisting completely. If an overlay crashes or
doesn't support Instant, we simply delete it and hide the dropdown. If a
preloaded NTP crashes or doesn't support Instant, we just delete it. In both
cases, we will fall back to the baked-in Instant page if necessary, and then
retry the remote page only on a user action.

BUG=233616

Review URL: https://chromiumcodereview.appspot.com/13873010

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@197864 4ff67af0-8c30-449e-8e8b-ad334ec8d88c