Files.app: Refresh the tree even when the top-level directory is changed.
When the directory structure is changed, the tree should reflect it immediately.

Previously, when a non-top level directory is changed, the list is changed as well. But when a top-directory is changed, the list is not changed. This is because a bug of the recursive strategy in DirectoryTreeUtil.updateChangedDirectoryItem(). This patch fixes it.

BUG=235107
TEST=manual

Review URL: https://chromiumcodereview.appspot.com/14677005

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@197882 4ff67af0-8c30-449e-8e8b-ad334ec8d88c