Files.app: Fixed a position of icon in the tasks menu
Originally, the icon of tasks menu item is located at the right of menu and
no space is inserted for the icon.
This CL moves the icon to the left edge of menu and inserts space in order not to overlap the text label and the icon.

BUG=237157
TEST=Open tasks menu and check the position of icon

Review URL: https://chromiumcodereview.appspot.com/14846005

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@197919 4ff67af0-8c30-449e-8e8b-ad334ec8d88c