Fix: 'You have no extensions' message overlapping installed extensions.
If the ID specified as a query parameter doesn't match any extension in the list, there is a JS error when calculating the offset of the highlight, which leads to the message not being hidden. This fixes the JS error so the execution flow can continue normally.

BUG=224191

Review URL: https://chromiumcodereview.appspot.com/14852008

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@197971 4ff67af0-8c30-449e-8e8b-ad334ec8d88c