Pass the hashed username to the NetworkConfigurationUpdater.
This change also simplifies the parameters that are passed to the
NetworkConfigurationUpdater when user policy is ready. Eventually a
NetworkConfigurationUpdater will be created for each Profile, and these
parameters will be passed directly in the constructor.

The hashed username will be used in a subsequent CL (https://codereview.chromium.org/13957012).

BUG=157696

Review URL: https://chromiumcodereview.appspot.com/14771010

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@198398 4ff67af0-8c30-449e-8e8b-ad334ec8d88c