InstantExtended: don't reset InstantTab if not ready.
Currently, we reset InstantTab if it's not ready by the time user starts
typing. With this change, we switch over to using the overlay, but leave the
tab connected so it has a chacne to finish loading. We'll try the tab again
when the overlay is hidden or committed.

BUG=236594

Review URL: https://chromiumcodereview.appspot.com/14843002

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@198460 4ff67af0-8c30-449e-8e8b-ad334ec8d88c