This is the first of two patches to drag and drop items from the app list to the launcher.
Everything basically works with this patch, but two essential things are still missing:
1. The icon which gets dragged should get its own widget so that it can get visually dragged outside the app list.
2. The unit tests.

They will be send as a second patch because of:
a. The patch is already pretty big as it is.
b. I want to make get this "signed off" before continuing this route.

BUG=166429
TEST=visual, tests come with second patch

Review URL: https://chromiumcodereview.appspot.com/14533006

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@199016 4ff67af0-8c30-449e-8e8b-ad334ec8d88c