Open a browser tab when Ash on Windows is launched.
The current behavior is to launch ash with nothing open
so the user would need to click on the Chrome ash icon
to see the browser.

This change creates two new notification types which
we use to show the Ash browser window. It is done this
way to preserve symmetry with
chrome_shell_delegate_chromeos.cc 

BUG=227247
TEST=open ash a window should be shown.

Review URL: https://chromiumcodereview.appspot.com/15021003

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@199049 4ff67af0-8c30-449e-8e8b-ad334ec8d88c