Added 'startDragSelection' funciton to cr.list.
This function shows the border of drag selection and adds mouse event handlers 
so that the border can be resized and cleared by mouse operations. 

This function will be called from handlers of 'dragstart' event, that are 
added from the outside of cr.list. If 'startDragSelection' is called from the 
handlers, drag selection is started instead of normal item dragging. 

In this CL, the actual behavior of selecting items is not implemented and 
it is planned to be implemented in the next CL. 

https://docs.google.com/file/d/0B4oFnlAOOE2CRTJzYjJBc1Jtenc/edit?usp=sharing

BUG=224832
TEST=Adds a line 'list.startDragSelection(event); return;' to the beginning of
     FileTransferController#onDragStart_, opens the file manager, drags on the
     file list.

Review URL: https://chromiumcodereview.appspot.com/14493013

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@199181 4ff67af0-8c30-449e-8e8b-ad334ec8d88c