Identity API: Copy WebAuthFlow -> ExperimentalWebAuthFlow
The behavior of identity.launchWebAuthFlow and 
experimental.identityLaunchWebAuthFlow is about to diverge. This 
change splits the WebAuthFlow class into experimental and 
non-experimental versions in advance of the divergence. 

The API tests are also changing to use net::SpawnedTestServer to 
avoid data: URL hacks that won't work in the future implementation. 

BUG=228908

Review URL: https://chromiumcodereview.appspot.com/14828005

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@199191 4ff67af0-8c30-449e-8e8b-ad334ec8d88c