Override server-side simple-cache trial with commandline switches.
FIXING broken about:flags for simple cache on Android.

Until recently we used a trick to avoid dedicated global settings in
net::disk_cache other than the objects belonging to the infrastructure of the
trials. The commandline flag would change the experiment group on the chromium
side, while one the net/ side we would check which group we belong to.

Later we decided to switch to the server-controlled experiments. This went into
a conflict with our approach: the server-controlled experiment would override
the settings set by the command-line option. This is unfortunate.

Stevet@ suggests that there are plans to avoid overriding the experiment groups
by the server config, but in the meantime the trick stops working.

Solve this by adding backend type enum provided at backend instantiation.

TBR=jam@chromium.org
BUG=239461

Review URL: https://chromiumcodereview.appspot.com/14828004

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@199708 4ff67af0-8c30-449e-8e8b-ad334ec8d88c