Stop reacting to profiles being destroyed when activating profiles.
Currently we prevent posting multiple times by stopping observing for the
window ready notification but MaybeActivateProfile can be called by the profile
destruction as well. This leads to cases where we try to remove the window
ready notification multiple times although it was already removed.

This patch stops listening for profile destruction as well in that case.

BUG=229343
R=mad@chromium.org, pkasting@chromium.org

Review URL: https://chromiumcodereview.appspot.com/14633006

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@199710 4ff67af0-8c30-449e-8e8b-ad334ec8d88c