Translate: Add UMAs to know usages
Translate.InitiateStatus is an enum histogram to know what is the reason of
Chrome Translate behaviors when it's ready to perform translation.
It is good resource to know how many people use Chrome Translate, disable,
doesn't visit lages in foreign language, and so on.

BUG=239711
TEST=unit_tests --gtest_filter='TranslateManagerMetricsTest.*'
R=mad@chromium.org

Review URL: https://codereview.chromium.org/15094005

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@199712 4ff67af0-8c30-449e-8e8b-ad334ec8d88c