[Downloads] Use Resume() instead of ResumeInterruptedDownload()
DownloadItem had both a Resume() method and a
ResumeInterruptedDownload() method. The former was used for paused (and
potentially interrupted) downloads while the latter was used for
interrupted downloads.

This change makes ResumeInterruptedDownload() internal to
DownloadItemImpl and only exposes Resume() as a public method on
DownloadItem. This should reduce confusion in the API. Resume() also
correctly resets the auto-resume count for a DownloadItem.

BUG=7648

Review URL: https://chromiumcodereview.appspot.com/14954002

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@200061 4ff67af0-8c30-449e-8e8b-ad334ec8d88c