[Storage Monitor] Check for floppy drives with both X: and \\.\X: forms of device name.
We observe that in Windows 7 and 8, it looks like using the "X:" device name for
QueryDosDevice doesn't work -- just returns "/". Using the slash-ified name is working.
We aren't sure whether this does or does not work for older windows versions yet, so
for now, trying both variants.

R=vandebo@chromium.org
BUG=234318

Review URL: https://chromiumcodereview.appspot.com/14576016

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@200195 4ff67af0-8c30-449e-8e8b-ad334ec8d88c