Remove google_apis::UploadMode.
It was used to use for checking the http method, but now the code has been
gone, and the enum value is only used for sanity check of succeess response
code. We can simplify the response code handling, as it is done in
DriveUploader.
To return the correct response code from FakeDriveService, InitiateUpload methods will embed an original query parameter to upload_url.

BUG=240996
TEST=Ran unit_tests

Review URL: https://chromiumcodereview.appspot.com/14748008

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@200359 4ff67af0-8c30-449e-8e8b-ad334ec8d88c