Make the constants in defaults.cc match their declaration order in defaults.h.
This also reorders a couple constants in the header so that the .cc file can
combine more groups of constants under fewer #ifdefs.  The exception to this is
the "autocomplete font size" constants block, which will be getting mostly nuked
in the future so I'm not worried about combining it with anything.

Should be no changes in constant values.

BUG=none
TEST=none
TBR=sky@chromium.org

Review URL: https://codereview.chromium.org/15208002

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@200383 4ff67af0-8c30-449e-8e8b-ad334ec8d88c