cc: Remove predictor benchmark
The predictor benchmark is currently not correct.
However, it's also not used. It's not a part of
telemetry, or UMA. I'd like to remove it in hopes
of cleaning up the code. If we do want this measurement
I propose we find a good spot for it so that it doesn't
add too much code complexity.

BUG=240500

Review URL: https://chromiumcodereview.appspot.com/14587011

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@200557 4ff67af0-8c30-449e-8e8b-ad334ec8d88c