[InfoBar] Add InfoBarDelegate::GetIconID()
Splits apart the GetIcon() call to allow accessing the ID of the resource in
addition to the loaded image.  This is needed so that the Android port can
determine which of the Java-side resources it needs to load for its UI.

* GetIconID() is a function that gets overridden by subclasses instead of
  GetIcon().  GetIconID() returns the IDR_* tag that represents the icon that
  it would normally load.

* GetIcon() now calls GetIconID() and shouldn't be overridden.

* Small change to the SimpleAlertInfoBarDelegate::Create() call to account
for invalid icon IDs (was NULL for bitmaps, 0 for non-existent icons).

BUG=237034

Review URL: https://chromiumcodereview.appspot.com/15067008

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@200598 4ff67af0-8c30-449e-8e8b-ad334ec8d88c