Add interfaces for resuming upload process terminated before completion.
This CL adds ResumeUploadFile method to DriveUploaderInterface, and
upload_location to the UploadCompletionCallback to use it.
Instead, this CL removes drive_file_path and local_file_path from
UploadCompletionCallback, which are little used now.
The implementation of ResumeUploadFile will be done in following CLs.

BUG=240565
TEST=Ran unit_tests

Review URL: https://chromiumcodereview.appspot.com/14875012

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@200647 4ff67af0-8c30-449e-8e8b-ad334ec8d88c