Do not set session preference to LAST for new profiles
Browser restarts that should restore sessions (such as after changing a flag)
launch the new browser process with a special flag, which is then used to set
the session preference to LAST.

The problem is that this flag now stays active for that browser process, so
creating a new profile from the relauched process will result in the session
preference being wrongfully set to LAST in the new profile, which stops the
new profile from loading the appropriate startup URLs.

A check for new profiles was added so that the preference is no longer set for them.

BUG=229343
R=pkasting@chromium.org

Review URL: https://chromiumcodereview.appspot.com/14004005

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@200782 4ff67af0-8c30-449e-8e8b-ad334ec8d88c